@@ -1,17 +1,15 @@
 import MainPageCarousel from "@/components/molecules/MainPageCarousel";
 
-import Image from 'next/image'
-import {Button} from "@/components/ui/button";
+import Image from "next/image";
+import { Button } from "@/components/ui/button";
 import React from "react";
 import PopularWines from "@/components/molecules/PopularWines";
 import Link from "next/link";
 
 import Section from "@/components/templates/Section";
-import AvisSimple from "@/components/molecules/AvisSimple";
-import {MainPageEvents} from "@/components/molecules/EventsList";
+import { MainPageEvents } from "@/components/molecules/EventsList";
 
 export default function Home() {
-
   return (
     <main>
       <Section id={`presentation`}>
@@ -19,21 +17,35 @@ export default function Home() {
           className={`flex flex-col gap-4 px-20 pb-10 text-center w-full text-primary`}
         >
           <h1 className={`flex justify-center text-6xl md:text-8xl`}>
-            <p className={`bg-clip-text text-transparent bg-gradient-to-r from-green-900 to-green-200 `}>Minca</p>
+            <p
+              className={`bg-clip-text text-transparent bg-gradient-to-r from-green-900 to-green-200 `}
+            >
+              Minca
+            </p>
           </h1>
           <p className={`text-3xl md:text-6xl`}>Vinhos Naturais</p>
         </div>
-        <div className={`flex flex-col lg:flex-row gap-20 justify-center items-center`}>
+        <div
+          className={`flex flex-col lg:flex-row gap-20 justify-center items-center`}
+        >
           <div
             className={`flex flex-col justify-between items-center w-[350px] md:w-[400px] h-[350px] md:h-[380px]`}
           >
             <div
               className={`
             flex flex-col items-center justify-between
-            w-full h-full bg-[#F6E9D9] rounded-xl shadow-[inset_0_2px_5px_1px_rgba(0,0,0,5)] p-5`}>
-              <Image priority={true} alt={`Mincas`} src={`/Logo2.png`} width={200} height={200} className={`w-[60%] h-auto`}/>
+            w-full h-full bg-[#F6E9D9] rounded-xl shadow-[inset_0_2px_5px_1px_rgba(0,0,0,5)] p-5`}
+            >
+              <Image
+                priority={true}
+                alt={`Mincas`}
+                src={`/Logo2.png`}
+                width={200}
+                height={200}
+                className={`w-[60%] h-auto`}
+              />
               <div
-                className={`flex flex-col items-center gap-5 
+                className={`flex flex-col items-center gap-5
                 font-medium text-lg`}
               >
                 <p className={`text-center`}>
@@ -44,14 +56,19 @@ export default function Home() {
                 </p>
               </div>
               <div className={`grid grid-cols-2 gap-5`}>
-                <Link href={`/wines`}><Button>Nossos Vinhos</Button></Link>
-                <Link href={`/about`}><Button>Sobre Nos</Button></Link>
+                <Link href={`/wines`}>
+                  <Button>Nossos Vinhos</Button>
+                </Link>
+                <Link href={`/about`}>
+                  <Button>Sobre Nos</Button>
+                </Link>
               </div>
             </div>
           </div>
           <div
-            className={`xl:row-start-1 flex justify-center items-center w-[350px] md:w-[400px] h-[350px] md:h-[380px]`}>
-            <MainPageCarousel/>
+            className={`xl:row-start-1 flex justify-center items-center w-[350px] md:w-[400px] h-[350px] md:h-[380px]`}
+          >
+            <MainPageCarousel />
           </div>
         </div>
       </Section>
@@ -59,48 +76,72 @@ export default function Home() {
         <div
           className={`py-16 gap-8 flex flex-col items-center justify-center`}
         >
-          <PopularWines/>
+          <PopularWines />
           <Link href={`/wines`}>
-            <Button>
-              Ver Mais
-            </Button>
+            <Button>Ver Mais</Button>
           </Link>
         </div>
       </Section>
       <Section id={`about`}>
         <div className="max-w-4xl mx-auto px-4 py-16">
-          <h2 className="text-3xl font-bold text-primary dark:text-secondary mb-8">Sobre Vinhos Naturais</h2>
+          <h2 className="text-3xl font-bold text-primary dark:text-secondary mb-8">
+            Sobre Vinhos Naturais
+          </h2>
           <div className="grid md:grid-cols-3 items-center gap-8 p-3 rounded-lg relative">
             <div className="col-span-2">
               <p className="">
-                Vinhos naturais são produzidos com <span className="font-semibold text-primary">mínima intervenção</span>, do cultivo
-                das uvas ao engarrafamento, priorizando <span
-                className="text-primary">práticas sustentáveis e orgânicas</span>.
+                Vinhos naturais são produzidos com{" "}
+                <span className="font-semibold text-primary">
+                  mínima intervenção
+                </span>
+                , do cultivo das uvas ao engarrafamento, priorizando{" "}
+                <span className="text-primary">
+                  práticas sustentáveis e orgânicas
+                </span>
+                .
               </p>
-              <br/>
+              <br />
               <p className="font-bold text-lg">
-                Na <span className="bg-clip-text text-transparent bg-gradient-to-r from-green-800 to-green-400">Minca</span>, nossos vinhos naturais são:
+                Na{" "}
+                <span className="bg-clip-text text-transparent bg-gradient-to-r from-green-800 to-green-400">
+                  Minca
+                </span>
+                , nossos vinhos naturais são:
               </p>
               <ul className="list-disc pl-6 space-y-2 text-pretty">
                 <li className="">
-                  Feitos com <span className="text-primary underline">uvas cultivadas organicamente</span>, sem químicos
-                  sintéticos
+                  Feitos com{" "}
+                  <span className="text-primary underline">
+                    uvas cultivadas organicamente
+                  </span>
+                  , sem químicos sintéticos
                 </li>
                 <li className="">
-                  Fermentados com <span className="text-primary underline">leveduras naturais</span>
+                  Fermentados com{" "}
+                  <span className="text-primary underline">
+                    leveduras naturais
+                  </span>
                 </li>
                 <li className="">
-                  <span className="bg-clip-text text-transparent bg-primary">Livres de aditivos</span> e conservantes
-                  artificiais
+                  <span className="bg-clip-text text-transparent bg-primary">
+                    Livres de aditivos
+                  </span>{" "}
+                  e conservantes artificiais
                 </li>
                 <li className="">
-                  Não filtrados, preservando <span className="text-primary">sabores e nutrientes</span>
+                  Não filtrados, preservando{" "}
+                  <span className="text-primary">sabores e nutrientes</span>
                 </li>
               </ul>
-              <br/>
+              <br />
               <p className="font-bold text-lg">
-                Escolher os vinhos naturais da <span className="bg-clip-text text-transparent bg-primary">Minca</span> é
-                experimentar algo único e apoiar uma vinicultura <span className="text-primary">ecologicamente responsável</span>.
+                Escolher os vinhos naturais da{" "}
+                <span className="bg-clip-text text-transparent bg-primary">
+                  Minca
+                </span>{" "}
+                é experimentar algo único e apoiar uma vinicultura{" "}
+                <span className="text-primary">ecologicamente responsável</span>
+                .
               </p>
             </div>
             <div className={`hidden z-10 md:block relative`}>
@@ -128,16 +169,18 @@ export default function Home() {
       */}
 
       <Section bgColor={`primary`} id={`events`}>
-        <h2 className="text-3xl font-bold mb-8 md:pl-6 w-full flex justify-center">Nossos Vinhos Especiais</h2>
-        <>
-          Special wines
-        </>
+        <h2 className="text-3xl font-bold mb-8 md:pl-6 w-full flex justify-center">
+          Nossos Vinhos Especiais
+        </h2>
+        <>Special wines</>
       </Section>
 
       <Section id={`events`}>
-        <h2 className="text-3xl text-primary text-center md:text-left font-bold mb-8 md:pl-6 w-full">Próximos Eventos</h2>
+        <h2 className="text-3xl text-primary text-center md:text-left font-bold mb-8 md:pl-6 w-full">
+          Próximos Eventos
+        </h2>
         <div className="xl:flex grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 flex-1 gap-5 px-6">
-          <MainPageEvents/>
+          <MainPageEvents />
         </div>
       </Section>
     </main>
