@@ -1,25 +1,28 @@
-'use client'
+"use client";
 
-import { useEffect, useState } from 'react'
-import { useSearchParams } from 'next/navigation'
+import React, { useEffect, useState } from "react";
+import { useSearchParams } from "next/navigation";
 
 export default function ErrorMessage() {
-  const [errorMessage, setErrorMessage] = useState<string | null>(null)
-  const searchParams = useSearchParams()
+  const [errorMessage, setErrorMessage] = useState<string | null>(null);
+  const searchParams = useSearchParams();
 
   useEffect(() => {
-    const error = searchParams.get('error')
+    const error = searchParams.get("error");
     if (error) {
-      setErrorMessage(decodeURIComponent(error))
+      setErrorMessage(decodeURIComponent(error));
     }
-  }, [searchParams])
+  }, [searchParams]);
 
-  if (!errorMessage) return null
+  if (!errorMessage) return null;
 
   return (
-    <div className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative mb-4" role="alert">
+    <div
+      className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative mb-4"
+      role="alert"
+    >
       <strong className="font-bold">Error: </strong>
-        <span className="block sm:inline">{errorMessage}</span>
+      <span className="block sm:inline">{errorMessage}</span>
     </div>
-  )
+  );
 }
