@@ -1,3 +1,5 @@
+import React from "react";
+
 export default function ErrorPage() {
-  return <p>Sorry, something went wrong</p>
+  return <p>Sorry, something went wrong</p>;
 }
