@@ -1,38 +1,13 @@
-import type { Metadata } from "next";
-import "@/app/globals.css"
-import { Allerta_Stencil } from "next/font/google";
+import React from "react";
 
-export const metadata: Metadata = {
-  title: "Minca",
-  description: "Descobra nossa paixão para os vinhos naturais.",
-};
-
-const inter = Allerta_Stencil({
-  subsets: ["latin"],
-  weight: "400",
-});
-
-type LayoutProps = {
+export default function Layout({
+  children,
+}: {
   children: React.ReactNode;
-  isAuthPage?: boolean;
-};
-
-export default function CombinedLayout(
-  { 
-    children, 
-  }: 
-  {
-    children: React.ReactNode,
-  }) {
+}) {
   return (
-    <html lang="en">
-      <body className={`${inter.className}`}>
-          <div className="flex min-h-screen bg-background text-neutral text-md">
-            <div className="w-full">
-              {children}
-            </div>
-          </div>
-      </body>
-    </html>
+    <div className="flex min-h-screen bg-background text-neutral text-md">
+      <div className="w-full">{children}</div>
+    </div>
   );
 }
