@@ -1,5 +1,5 @@
+import React from "react";
+
 export default function Loading() {
-	return (
-			<div className="lds-dual-ring"/>
-	);
+  return <div className="lds-dual-ring" />;
 }
