@@ -1,11 +1,9 @@
+import React from "react";
+
 export const VerticalSeparator = () => {
-	return (
-		<div className={`w-0.5 h-full bg-primary mx-1 rounded-full`}></div>
-	)
-}
+  return <div className={`w-0.5 h-full bg-primary mx-1 rounded-full`}></div>;
+};
 
 export const HorizontalSeparator = () => {
-	return (
-		<div className={`w-full h-[4px] bg-primary my-2 rounded-full`}></div>
-	)
-}
+  return <div className={`w-full h-[4px] bg-primary my-2 rounded-full`}></div>;
+};
