@@ -1,3 +1,7 @@
-export function price (price: number) {
-	return `R$ ${price} `;
+export function price(price: number) {
+  return `R$ ${price} `;
+}
+
+export function formatDate(date: string) {
+  return new Date(date).toLocaleDateString("pt-BR");
 }
